lilium_api/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.1.4
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.10
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.1
